Move has_post
